feat: add product table
Write a program that multiplies the contents of two registers that you have initialized
using an immediate operand with the ori instruction. Determine (by inspection) the
number of significant bits in each of the following numbers, represented in two’s complement.
Use the program to form their product and then determine the number of significant bits
in it.